refactor objects to packages
