+++ src/main/java/com/odyssey/odyssey/domain/account/Account.java
@@ -1,5 +1,6 @@
-package com.odyssey.odyssey.domain;
+package com.odyssey.odyssey.domain.account;
 
+import com.odyssey.odyssey.domain.DebitCard;
 import java.math.BigDecimal;
 
 public class Account {
