+++ src/main/java/com/odyssey/odyssey/domain/account/AccountType.java
@@ -1,4 +1,4 @@
-package com.odyssey.odyssey.domain;
+package com.odyssey.odyssey.domain.account;
 
 public enum AccountType {
     CREDIT, DEBIT
