+++ src/main/java/com/odyssey/odyssey/domain/account/LoginInformation.java
@@ -1,4 +1,4 @@
-package com.odyssey.odyssey.domain;
+package com.odyssey.odyssey.domain.account;
 
 public class LoginInformation {
